@@ -13,7 +13,7 @@
         }
     ],
     "require": {
-        "symfony/symfony": "~2.5",
+        "symfony/symfony": "~2.7",
         "doctrine/orm": "~2.4,<2.5",
         "symbio/orangegate4-admin-bundle": "~1.0@dev"
     },
