@@ -32,6 +32,7 @@ class SystemLanguageTokenAdmin extends BaseAdmin
                 ->add('token', 'text', array('label' => 'Key', 'data' => $site.'.'));
         }
         $formMapper
+            ->add('catalogue', 'sonata_type_model_list', array())
             ->add('translations', 'sonata_type_collection', array(), array(
                 'edit' => 'inline',
                 'inline' => 'table',
@@ -45,6 +46,7 @@ class SystemLanguageTokenAdmin extends BaseAdmin
     {
         $datagridMapper
             ->add('token')
+            ->add('catalogue')
         ;
     }
 
@@ -62,6 +64,7 @@ class SystemLanguageTokenAdmin extends BaseAdmin
     {
         $listMapper
             ->addIdentifier('token')
+            ->addIdentifier('catalogue')
             ->add('_action', 'actions', array(
                 'actions' => array(
                     'edit' => array(),
@@ -77,51 +80,27 @@ class SystemLanguageTokenAdmin extends BaseAdmin
         $showMapper
             ->add('id')
             ->add('token', null, array('label' => 'Key'))
+            ->add('catalogue', null, array('label' => 'Catalogue'))
         ;
     }
 
     public function prePersist($object)
     {
-        $em = $this->modelManager->getEntityManager('SymbioOrangeGateTranslationBundle:LanguageToken');
-
-        $container = $this->getConfigurationPool()->getContainer();
-        $cacheDir = $container->get('kernel')->getCacheDir();
-        $finder = new \Symfony\Component\Finder\Finder();
-        $finder->in(array($cacheDir . "/../*/translations"))->files();
-
-        foreach($finder as $file){
-            unlink($file->getRealpath());
-        }
-
-        if (is_dir($cacheDir.'/translations')) {
-            rmdir($cacheDir.'/translations');
-        }
-
         foreach ($object->getTranslations() as $tr) {
             $tr->setLanguageToken($object);
         }
+
+        $this->clearCache();
     }
 
     public function preUpdate($object)
     {
-        $em = $this->modelManager->getEntityManager('SymbioOrangeGateTranslationBundle:LanguageToken');
+        $this->prePersist($object);
+    }
 
-        $container = $this->getConfigurationPool()->getContainer();
-        $cacheDir = $container->get('kernel')->getCacheDir();
-        $finder = new \Symfony\Component\Finder\Finder();
-        $finder->in(array($cacheDir . "/../*/translations"))->files();
-
-        foreach($finder as $file){
-            unlink($file->getRealpath());
-        }
-
-        if (is_dir($cacheDir.'/translations')) {
-            rmdir($cacheDir.'/translations');
-        }
-
-        foreach ($object->getTranslations() as $tr) {
-            $tr->setLanguageToken($object);
-        }
+    public function postRemove($object)
+    {
+        $this->clearCache();
     }
 
     /**
@@ -135,5 +114,21 @@ class SystemLanguageTokenAdmin extends BaseAdmin
         $query->andWhere($query->expr()->isNull($query->getRootAlias().'.site'));
 
         return $query;
+    }
+
+    public function clearCache()
+    {
+        $container = $this->getConfigurationPool()->getContainer();
+        $cacheDir = $container->get('kernel')->getCacheDir();
+        $finder = new \Symfony\Component\Finder\Finder();
+        $finder->in(array($cacheDir . "/../*/translations"))->files();
+
+        foreach($finder as $file){
+            unlink($file->getRealpath());
+        }
+
+        if (is_dir($cacheDir.'/translations')) {
+            rmdir($cacheDir.'/translations');
+        }
     }
 }
