@@ -2,9 +2,11 @@
 
 namespace Symbio\OrangeGate\TranslationBundle;
 
+use Symfony\Component\DependencyInjection\Compiler\PassConfig;
 use Symfony\Component\HttpKernel\Bundle\Bundle,
     Symfony\Component\DependencyInjection\ContainerBuilder,
-    Symbio\OrangeGate\TranslationBundle\DependencyInjection\Compiler\TemplatingPass;
+    Symbio\OrangeGate\TranslationBundle\DependencyInjection\Compiler\TemplatingPass,
+    Symbio\OrangeGate\TranslationBundle\DependencyInjection\Compiler\AddResourcePass;
 
 class SymbioOrangeGateTranslationBundle extends Bundle
 {
@@ -13,5 +15,6 @@ class SymbioOrangeGateTranslationBundle extends Bundle
         parent::build($container);
 
         $container->addCompilerPass(new TemplatingPass());
+        $container->addCompilerPass(new AddResourcePass(), PassConfig::TYPE_AFTER_REMOVING);
     }
 }
